\name{CompositionalDR-package}
\alias{CompositionalDR-package}
\docType{package}
\title{
Dimensionality Reduction with Compositional Data
}

\description{
Dimensionality Reduction with Compositional Data. Currently, Dirichlet component analysis and archetypal analysis are implemented.
}

\details{
\tabular{ll}{
Package: \tab CompositionalDR\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2025-11-26\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>
}

\references{
Wang H. Y., Yang Q., Qin H. and Zha H. (2008, July). Dirichlet component analysis: feature extraction for compositional data. In Proceedings of the 25th International Conference on Machine Learning (pp. 1128-1135).

Cutler A. and Breiman L. (1994). Archetypal analysis. Technometrics, 36(4): 338--347.
}

%\note{
%}


